fix crc16 to work with big size of data
git-svn-id: http://svn2.assembla.com/svn/secu3man@222 c1dae7de-a428-49f4-85d8-50b65f41d01d

+++ sources/io-core/CRC16.cpp
@@ -24,7 +24,7 @@
 
 #define      P_16        0xA001   
 
-unsigned short CRC_API crc16(const unsigned char *buf, unsigned short num)
+unsigned short CRC_API crc16(const unsigned char *buf, unsigned long num)
 {
 unsigned int i;
 unsigned short crc = 0xffff;
